test(modx): drop hollow live tests that assert MODX/testbench
Keep only MiniShop3-backed checks: phinx.php vs live dbname, processors,
models, DI, events. Leave the processors_path warning in the suite README.
